pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "credit-risk-model"
version = "0.1.0"
description = "A project for predicting credit risk on the German Credit dataset using a 4 model voting ensemble."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core data stack
    "numpy>=1.24.0,<3.0.0",
    "pandas>=2.0.0,<3.0.0",
    "scipy>=1.10.0,<2.0.0",
    "scikit-learn>=1.3.0,<2.0.0",
    # Model training
    "scikit-optimize>=0.10.0", # BayesSearchCV for LRC, RFC, SVC
    "category-encoders>=2.6.0", # WOE, Target, Count, Ordinal encoders
    "imbalanced-learn>=0.11.0", # SMOTE, SVMSMOTE
    "catboost>=1.2.0", # 4th ensemble model (native categorical handling)
    # Config + serialization
    "pydantic>=2.5.0,<3.0.0",
    "pydantic-settings>=2.1.0,<3.0.0",
    "pyyaml>=6.0.0,<7.0.0",
    "joblib>=1.3.0,<2.0.0",
    # Experiment tracking
    "mlflow>=2.9.0,<3.0.0",
    # Visualization (for MLflow artifact logging)
    "matplotlib>=3.7.0,<4.0.0",
    "seaborn>=0.12.0,<1.0.0",
    ]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-cov>=4.1.0,<6.0.0",
    "ruff>=0.3.0",
    "nox>=2024.3.2",
    "ipykernel>=6.25.0",
    "jupyter>=1.0.0",
    ]
app = [
    "streamlit>=1.28.0",
    ]

[tool.hatch.build.targets.wheel]
packages = ["src/credit_risk_model"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
